fix(rtmp): module structure
