--- crates/rtmp/src/command_messages/define.rs
@@ -1,35 +1,0 @@
-use std::borrow::Cow;
-
-use super::netconnection::NetConnectionCommand;
-use super::netstream::NetStreamCommand;
-use super::on_status::OnStatus;
-
-#[derive(Debug, Clone)]
-pub struct Command<'a> {
-    pub transaction_id: f64,
-    pub command_type: CommandType<'a>,
-}
-
-/// This enum wraps the [`NetConnectionCommand`] and [`NetStreamCommand`] enums.
-#[derive(Debug, Clone)]
-pub enum CommandType<'a> {
-    /// NetConnection command
-    NetConnection(NetConnectionCommand<'a>),
-    /// NetStream command
-    NetStream(NetStreamCommand<'a>),
-    /// onStatus command
-    OnStatus(OnStatus<'a>),
-    /// Any unknown command
-    ///
-    /// e.g. FFmpeg sends some commands that don't appear in any spec, so we need to handle them.
-    Unknown { command_name: Cow<'a, str> },
-}
-
-/// NetStream onStatus level (7.2.2.) and NetConnection connect result level (7.2.1.1.)
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum CommandResultLevel {
-    Warning,
-    Status,
-    Error,
-    Unknown(String),
-}
