+++ crates/rtmp/src/command_messages/mod.rs
@@ -1,10 +1,42 @@
-mod define;
-mod error;
+use std::borrow::Cow;
+
+use netconnection::NetConnectionCommand;
+use netstream::NetStreamCommand;
+use on_status::OnStatus;
+
+pub mod error;
 pub mod netconnection;
 pub mod netstream;
 pub mod on_status;
-mod reader;
-mod writer;
+pub mod reader;
+pub mod writer;
 
-pub use define::{CommandType, *};
-pub use error::CommandError;
+#[derive(Debug, Clone)]
+pub struct Command<'a> {
+    pub transaction_id: f64,
+    pub command_type: CommandType<'a>,
+}
+
+/// This enum wraps the [`NetConnectionCommand`] and [`NetStreamCommand`] enums.
+#[derive(Debug, Clone)]
+pub enum CommandType<'a> {
+    /// NetConnection command
+    NetConnection(NetConnectionCommand<'a>),
+    /// NetStream command
+    NetStream(NetStreamCommand<'a>),
+    /// onStatus command
+    OnStatus(OnStatus<'a>),
+    /// Any unknown command
+    ///
+    /// e.g. FFmpeg sends some commands that don't appear in any spec, so we need to handle them.
+    Unknown { command_name: Cow<'a, str> },
+}
+
+/// NetStream onStatus level (7.2.2.) and NetConnection connect result level (7.2.1.1.)
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum CommandResultLevel {
+    Warning,
+    Status,
+    Error,
+    Unknown(String),
+}
