+++ crates/rtmp/src/session/mod.rs
@@ -1,7 +1,5 @@
-mod error;
-mod handler;
+pub mod error;
+pub mod handler;
 mod server_session;
 
-pub use self::error::SessionError;
-pub use self::handler::{SessionData, SessionHandler};
-pub use self::server_session::Session;
+pub use server_session::ServerSession;
