+++ crates/rtmp/src/user_control_messages/mod.rs
@@ -1,4 +1,17 @@
-mod define;
-mod writer;
+pub mod writer;
 
-pub use define::EventMessageStreamBegin;
+nutype_enum::nutype_enum! {
+    pub enum EventType(u16) {
+        StreamBegin = 0,
+        StreamEOF = 1,
+        StreamDry = 2,
+        SetBufferLength = 3,
+        StreamIsRecorded = 4,
+        PingRequest = 6,
+        PingResponse = 7,
+    }
+}
+
+pub struct EventMessageStreamBegin {
+    pub stream_id: u32,
+}
